+++ CryptoTrader/Trading/Traders/Simple/Strategies/BTS/Models/TraderBTSTrade.swift
@@ -2,12 +2,12 @@ import Foundation
 
 struct TraderBTSTrade: Codable {
     
-    var price: Double
-    var quantity: Double
-    var value: Double
+    var price: Decimal
+    var quantity: Decimal
+    var value: Decimal
     var date: Date
     
-    init(price: Double, quantity: Double, value: Double, now: Date) {
+    init(price: Decimal, quantity: Decimal, value: Decimal, now: Date) {
         self.date = now
         self.price = price
         self.quantity = quantity
